Add alpha design of the player
Add:
- play, pause, paramter, audio, subtitle, fullscreen button
- definition 16:9 for the video
- dropdown autio, video, parameter
- hover the toolbar on the video container
- gray out the video container when the player is buffering
- tooltips on the button
- progress bar depending on the current time in the video
- message of error when the path in wrong in vlc player

Update:
README.md -> add tpls for bootstrap ui for the tooltip

Rq Bugs:

To over the toolbar on the player we need to use :
windowless="true", but this option produce bugs. For example fullscreen
option doesn't work anymore.

The option audio and subtitle don't seem to work too.

Tested on Chrome and Firefox, Windows 7 64 bits