Fix fullscreen mode
Problem:
Button fullscreen doesn't work when windowless is enabled.

Fix:
Video is stopped and a new style is apply to the embed video, the the
video is played again.

Update:
CSS is updated, now the player no longer depends on CSS, only the toolbar. it's now a real
extern module !

Tested:
Chrome
Firefox

Rq bugs:

The escape button in fullscreen mode break the width and height of the
player, tested on Chrome and Firefox